remove compiler warnings

+++ parameter.h
@@ -83,8 +83,8 @@ public:
     int getNumberOfValues() const;
     std::vector<ParameterValue>& getValues();
     const std::vector<ParameterValue>& getValues() const;
-    void deleteValue(int idx);
-    void deleteTemperature(int idx);
+    void deleteValue(size_t idx);
+    void deleteTemperature(size_t idx);
     void importValues(const QString& filename);
 
     void sort();
